+++ NEWS.md
@@ -1,3 +1,25 @@
+# shinyB2AIDashboard 0.1.0
+
+## New functionality
+
+* Add function `dssrc_dataframe` to fetch b2ai program onboarding dataset
+  from google spreadsheet
+* Add function `orcid_api_call` as a simple client for the 
+  [ORCID API](https://pub.orcid.org/v3.0/#/Development_Public_API_v3.0).
+* Add function `works_from_orcid` to get all the publication 
+  records for a given ORCID.
+* Add function `orcid_work_detail` to fetch the details of a given
+  ORCID work record. 
+* A number of helper functions support the end-user functions above.
+
+## Other
+
+* Add documentation for new functions.
+* Add _pkgdown.yml base.
+* Add tests for `orcid_api_call`
+* Add tests for `works_from_orcid`
+* Basic pkgdown website published as gh-pages
+
 # shinyB2AIDashboard 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.